Big push before refactoring. Rack::Reshow is now in in alpha.
Huge commit, which I failed to separate. The bar that controls
viewing revisions is now added to all requests that return an object
that can be scanned (generally a string) and that contains <head>
and <body> tags (they're necessary for the sub!-ing that's used
to inject our html and css and javascript into the page).

Static files are served through a Rack::Static middleware. Much
easier than I expected, though the documentation is a bit rough
around the edges regarding where files are searched for when
defining a root dir.

The jQuery library that's injected is pulled from Google's servers.
What this means is less coding for me, but the middleware only works
when in an environment with an internet connection.

And... that's about it. Refactoring and updating tests comes next.